refactor(ffi): wrap EasyTierTcpStream for DataPlaneTcpStream; move dataplane logic to gateway::socks5::dataplane; box the guard of Socks5AutoConnector
